requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pillow>=10.0.0
mercadopago>=2.2.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
jinja2>=3.1.0
face-recognition>=1.3.0